pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "SEO analyzer"
authors = ["Frisson"]
readme = "README.md"
packages = [{include = "page_analyzer"}]
repositories = [{ url = "https://github.com/FrissonFrisson/python-project-83.git", type = "git" }]

[tool.poetry.dependencies]
python = ">=3.10, <4.0"
flask = "^3.0.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
validators = "^0.22.0"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
bs4 = "^0.0.1"
pip = "^23.3.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
